+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changes
 - Update package infrastructure to follow APE17 guidelines. [#142]
 - Added new functions ``healpix_to_xyz`` and ``xyz_to_healpix`` to
   convert to/from cartesian coordinates. [#141]
+- Add ``HEALPix.level`` property to the high-level interface.
+  This is a shortcut for the ``nside_to_level`` function. [#147]
 
 0.5 (2019-11-25)
 ================